pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "will-byers"
version = "0.1.0"
description = "Talk to Will Byers from Stranger Things through LED Christmas lights"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Kyle Johnson"},
]
dependencies = [
    "anthropic>=0.40.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "sounddevice>=0.4.0",
    "openai-whisper>=20230314",
    "rpi-ws281x>=5.0.0; platform_machine == 'aarch64' or platform_machine == 'armv7l'",
    "adafruit-circuitpython-neopixel>=6.0.0; platform_machine == 'aarch64' or platform_machine == 'armv7l'",
]

[project.scripts]
will-byers-chat = "will_byers.chat:main"
will-byers-voice-chat = "will_byers.voice_chat:main"
will-byers-create-mapping = "will_byers.create_mapping:main"

[project.urls]
Homepage = "https://github.com/kyjohnso/will_byers"

[tool.setuptools.packages.find]
where = ["."]
include = ["will_byers*"]
